+ explanation for existence
"Learn how to make clothes. Learn how to make clothes. Learn how to make clothes. Don’t do sketches and collages, because that’s bullshit I mean anybody can do that. You have to learn how to make clothes. And the other thing is, you have to shut up and work and work and work all the time, and produce and produce and produce, and the more you produce, your character or your talent will emerge. For better or for worse, your identity, your personality, your vision, whatever. As long as you work and work and work and make enough things to choose from, you’ll be able to have an edit, you’ll be able to edit something that becomes who you are. It’s so simple. You don’t have to go to parties and you don’t have to meet the right people, you just have to work and work and work and get your stuff out there."

— Rick Owens